changed structure of info
